Rename state to save data in store
The data for InGame, Heaven and Input were stored in 'states'.  But bevy
already uses the name state to distinguish between game states like
InGame, Pause, Heaven and so we would have the same name for two
different thing.

Now a singleton storage is called store instead of state to avoid
confusion.